Location report entrypoint
Moves the location report example to a package entrypoint.
