+++ .dockerignore
@@ -2,3 +2,4 @@
 *.pyc
 *.zip
 *.tar
+*requirements.txt
