+++ .travis.yml
@@ -7,10 +7,10 @@ python:
    - 3.7
 
 before_install:
-  - docker-compose build
+  - docker-compose build timemap-tests
 
 script:
-   - docker-compose run timemap pytest
+   - docker-compose run timemap-tests pytest
 
 deploy:
   provider: pypi
