+++ README.md
@@ -1,7 +1,5 @@
 # Timemap
 
-> repository under renovation
-
 A python package to deal with location history files, such
 as GMAPS takeout json.
 
@@ -22,21 +20,13 @@ or command line parameters (requires extension).
 
 # Configuration File
 The configuration consist of a yaml file. All the items in it, will be
-exposed as settings to your script.
+exposed as settings to your script. 
 
-
-``` yaml
-filepath: ./.history.json
-start: 2019-03-01
-end: 2019-03-31
-
-radius: 200 # in meters
-latitude: 44.434 # in decimal format
-longitude: 22.213 # in decimal format
-```
+Please see an example from [defaults.yaml][defaults.yaml].
 
 
 # Contributing
-Feel free to send me your pull requests
+Feel free to send me your pull requests and create issues for your ideas.
 
 
+[defaults.yaml]: ./tests/defaults.yaml
