+++ container/Dockerfile
@@ -1,4 +1,4 @@
-FROM python:3 as build
+FROM python:3
 
 WORKDIR /app
 
@@ -14,22 +14,10 @@ RUN pip install --no-cache-dir -r dev-requirements.txt
 COPY . .
 
 RUN pip install -e .
+
+CMD [ "timemap-report" ]
+
+FROM python:3 as build
+
 RUN ./.ci/build.sh
 RUN ./.ci/test.sh
-
-FROM python:3 
-
-WORKDIR /app
-
-COPY --from=build /app/requirements.txt .
-RUN pip install --no-cache-dir -r requirements.txt
-
-COPY --from=build /app/dist/*.whl . 
-COPY --from=build /app/tests/gmaps_sample.json .history.json 
-COPY --from=build /app/examples/*.py . 
-COPY --from=build /app/defaults.yml .
-
-RUN pip install --no-cache-dir *.whl && rm *.whl
-
-CMD [ "python", "location_report.py" ]
-
