+++ setup.py
@@ -65,4 +65,7 @@ setup(
     keywords=about["__keywords__"],
     packages=find_packages(exclude=["contrib", "docs", "tests", "examples"]),
     install_requires=get_requirements("requirements.txt"),
+    entry_points={
+        "console_scripts": ["timemap-report=timemap.__main__:location_report"]
+    },
 )
